final_project/requirements/base.txt
django==2.2.8
djangorestframework==3.10.3
drf-yasg==1.17.0
python-decouple==3.1
psycopg2==2.8.4
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.4.0
django-allauth==0.41.0